Move block context to a dedicated header
